requirements.txt
Click==7.0
elasticsearch==6.3.1
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mongo==0.2.0
pymongo==3.7.2
PyYAML==3.13
urllib3==1.24.1
Werkzeug==0.14.1